Register calendar before initialize
Client code can define hooks for fullCalendar in the options that can be called during initialization.  For example if a caller specifies "viewRender" and inside viewRender they would like to access the calendar object, this isn't possible since the uiCalendarConfig is not yet configured.  By adding the new calendar to uiCalendarConfig before calling fullCalendar(options) this allows this functionality.